% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim}
\alias{optim}
\title{Optimize}
\usage{
optim(xr, x, z0, steps, maxZ, maximize = 1L, verbose = 0L)
}
\arguments{
\item{xr}{The x-values not in the top-k.}

\item{x}{The x-values.}

\item{z0}{The vector of starting values for the optimization.}

\item{steps}{The number of optimization steps.}

\item{maxZ}{The maximum value of elements of z.}

\item{maximize}{Whether to maximize.}

\item{verbose}{Whether to report progress.}
}
\description{
Using the OLS equation for regression parameters, this function
maximizes or minimizes the regression slope subject to the following
conditions
1. The predictor is completely observed
2. Only the k-largest values of the target are observed
3. The sum of the entire target variable is observed

This function assumes that the domain of both x and z are non-negative 
integers excluding zero.

If the function is called with `maximum != 1`, it will return the solution
that minimizes the regression slope.
}
