% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_aggregate}
\alias{optim_aggregate}
\title{Optimize on Grouped Data}
\usage{
optim_aggregate(x_data, index, maximize = 1L, verbose = 0L)
}
\arguments{
\item{x_data}{See details.}

\item{index}{See details.}

\item{maximize}{Whether to maximize.}

\item{verbose}{Whether to report progress.}
}
\description{
A version of `optim()` designed to be used when the data are grouped.
}
\details{
This function optimizes the OLS regression slope subject to the same conditions as
`optim()`, but allows for the data to be grouped such that 
each group has its own optimization constraints and maximum number of 
optimization steps. The expectation is that the data represents cities grouped by state.

The \code{x_data} argument should be a dataframe containing city populations.
This data frame should have columns called \code{state}, the state in which
that city is located; \code{x}, the city population; and \code{isTopK},
`1` when that city is in the top-K and `0` otherwise.

The \code{index} object should be a dataframe containing the optimization
parameters for each state, to be used as a look-up table. This dataframe 
should have columns called \code{state}, \code{steps}, the number of recovered
firearms in that state; and \code{maxZ}, the maximum number of firearms that could be 
recovered in a non-top-K city.
}
