\name{topkOptim-package}
\alias{topkOptim-package}
\alias{topkOptim}
\docType{package}
\title{
  Top-K Optimization Functions.
}
\description{
  Optimization functions for top-k regression with a known total. This package contains
  functions to determine the maximum or minimum slope achievable when using OLS regression
  under the conditions that only k observations have known y-values, all x-values are
  known, and the total of observed and unobserved y-values are known.
}
\author{
Christopher Buglino, Ines Figueira

Maintainer: Christopher Buglino <cb5176@nyu.edu>
}
\keyword{ package }
